tspp: Vote for msm-bus clocks before enabling tspp clocks
Currently tspp driver votes for its core clocks, however, it is
possible that the source of the core-clocks will be off. When the
source is off, the tspp core is not clocked. This patch adds
voting for the msm-bus clock, thus guarantee clocking to
the tspp core.

The voting is done only when driver is asked to enable TSIF data
path to stream data through TSIF interface, at which the core clocks
are enabled as well.

Change-Id: I08cd25c0a23d1d69afa54cefddc0b2cb25b3dd7d
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>