coresight: add coresight fuse driver
Add support for CoreSight Fuse driver which can be used by other
CoreSight drivers to query the state of the CoreSight fuses to
determine if the Hardware they manage is functionally disabled or
not.

Drivers can then take necessary actions like failing the probe if
the Hardware they manage is functionally disabled.

Change-Id: I2c2a2af064db750539b0ffa33870ddb7a2832e68
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>