ASoC: msm: Change the definition of port_session variable
The definition of the post session variable used during port
mixer command is 32 bit wide, if the port number selected
is more than 32 bit and set_bit will end up in changing the
variable other than port_session as it is out of bound access.
Change the port session variable to 64 bit to take care of out of bound
access.

CRs-Fixed: 502025
Change-Id: I86b21d2dcfe6773b255dbf4a85eaa8f87e86826e
Signed-off-by: Santosh Mardi <gsantosh@codeaurora.org>

Conflicts:
	sound/soc/msm/qdsp6v2/msm-pcm-routing-v2.c